+++ src/server/mod.rs
@@ -5,6 +5,7 @@ use neva::App;
 use neva::types::ToolSchema;
 
 use crate::agent_id::{ROLE_EXECUTOR, ROLE_SUPERVISOR, agent_id};
+use crate::platform;
 
 mod prompts;
 mod resources;
@@ -17,8 +18,8 @@ pub enum Role {
 }
 
 pub async fn start(role: Option<Role>, agent_name: String, agent_index: u32) -> Result<()> {
-    set_serve_process_name();
-    install_serve_parent_lifecycle_hooks();
+    platform::set_serve_process_name();
+    platform::install_serve_parent_lifecycle_hooks();
 
     let role_str = match &role {
         Some(Role::Supervisor) => ROLE_SUPERVISOR,
@@ -119,41 +120,4 @@ pub async fn start(role: Option<Role>, agent_name: String, agent_index: u32) -> 
 
     app.run().await;
     Ok(())
-}
-
-fn set_serve_process_name() {
-    #[cfg(target_os = "linux")]
-    unsafe {
-        let name = b"ferrus-mcp\0";
-        let _ = libc::prctl(libc::PR_SET_NAME, name.as_ptr() as libc::c_ulong, 0, 0, 0);
-    }
-}
-
-fn install_serve_parent_lifecycle_hooks() {
-    #[cfg(target_os = "linux")]
-    {
-        unsafe {
-            let _ = libc::prctl(
-                libc::PR_SET_PDEATHSIG,
-                libc::SIGTERM as libc::c_ulong,
-                0,
-                0,
-                0,
-            );
-        }
-
-        let parent_pid = unsafe { libc::getppid() };
-        if parent_pid > 1 {
-            tokio::spawn(async move {
-                let mut interval = tokio::time::interval(std::time::Duration::from_secs(1));
-                loop {
-                    interval.tick().await;
-                    let current_ppid = unsafe { libc::getppid() };
-                    if current_ppid <= 1 || current_ppid != parent_pid {
-                        std::process::exit(0);
-                    }
-                }
-            });
-        }
-    }
 }
